extract decorations to separate file
